+++ main/src/main/java/net/oneandone/stool/cli/Main.java
@@ -101,8 +101,8 @@ public class Main {
                       cli.add(Port.class, "port port { port*(port) }");
                       cli.add(Refresh.class, "refresh -build=@refresh.build -restore");
                       cli.add(Remove.class, "remove -batch -force -backstage { setBackstage(backstage) }");
-                      cli.add(Restart.class, "restart");
-                      cli.add(Start.class, "start -tail");
+                      cli.add(Restart.class, "restart -nocache");
+                      cli.add(Start.class, "start -tail -nocache");
                       cli.add(Status.class, "status -defaults=@status.defaults info* { select*(info) }");
                       cli.add(Stop.class, "stop -sleep");
                       cli.add(Validate.class, "validate -email -repair");