+++ main/src/test/java/net/oneandone/stool/docker/EngineIT.java
@@ -35,7 +35,7 @@ public class EngineIT {
         String container;
 
         engine = open();
-        image = engine.imageBuild("sometag", Collections.emptyMap(), df("FROM debian:stretch-slim\nCMD ls -la /dev/fuse\n"), null);
+        image = engine.imageBuild("sometag", Collections.emptyMap(), df("FROM debian:stretch-slim\nCMD ls -la /dev/fuse\n"), false,null);
         container = engine.containerCreate(image, "somehost");
         engine.containerStart(container);
         Thread.sleep(1000);
@@ -54,7 +54,7 @@ public class EngineIT {
         labels = Strings.toMap("stooltest", UUID.randomUUID().toString());
         engine = open();
         assertTrue(engine.imageList(labels).isEmpty());
-        engine.imageBuild("sometag", labels, df("FROM debian:stretch-slim\nRUN touch abc\nCMD echo hello\n"), null);
+        engine.imageBuild("sometag", labels, df("FROM debian:stretch-slim\nRUN touch abc\nCMD echo hello\n"), false,null);
         ids = engine.imageList(labels);
         assertEquals(1, ids.size());
         image = ids.get(0);
@@ -80,7 +80,7 @@ public class EngineIT {
         message = UUID.randomUUID().toString();
 
         engine = open();
-        image = engine.imageBuild("sometag",  Collections.emptyMap(), df("FROM debian:stretch-slim\nCMD echo " + message + ";sleep 5\n"),null);
+        image = engine.imageBuild("sometag",  Collections.emptyMap(), df("FROM debian:stretch-slim\nCMD echo " + message + ";sleep 5\n"),false,null);
         assertNotNull(image);
 
         container = engine.containerCreate(image, "foo", false,
@@ -128,7 +128,7 @@ public class EngineIT {
         message = UUID.randomUUID().toString();
 
         engine = open();
-        image = engine.imageBuild("sometag",  Collections.emptyMap(), df("FROM debian:stretch-slim\nCMD echo " + message + ";sleep 5\n"),null);
+        image = engine.imageBuild("sometag",  Collections.emptyMap(), df("FROM debian:stretch-slim\nCMD echo " + message + ";sleep 5\n"), false,null);
         container = engine.containerCreate(image, "foo", false,
                 limit, null, null, Collections.emptyMap(), Collections.emptyMap(), Collections.emptyMap());
         engine.containerStart(container);
@@ -141,7 +141,7 @@ public class EngineIT {
         engine.imageRemove(image);
 
         engine = open();
-        image = engine.imageBuild("sometag",  Collections.emptyMap(), df("FROM debian:stretch-slim\nCMD echo " + message + ";sleep 5\n"),null);
+        image = engine.imageBuild("sometag",  Collections.emptyMap(), df("FROM debian:stretch-slim\nCMD echo " + message + ";sleep 5\n"), false,null);
         container = engine.containerCreate(image, "foo", false,
                 limit, null, null, Collections.emptyMap(), Collections.emptyMap(), Collections.emptyMap());
         engine.containerStart(container);